requirements.txt
viam-sdk>=0.60.0
spotipy>=2.23.0
aiohttp>=3.9.0
qrcode[pil]>=7.4.0
colorthief>=0.2.1
requests>=2.31.0
